Rename `page` folder to `routes`
